+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Rename `splitOnce` to `splitFirst`.
 * Add `splitTwice`.
 * Add `equalsLatin{1,2,3,4}`
+* Add `ifoldl'`.
 
 ## 0.1.4.0 -- 2019-11-12